tests/requirements.txt
pytest
Pillow
numpy
# Install this from the root directory to work
.